requirements.txt
PyYAML==3.12
requests==2.5.1
furl==0.5.6
terminaltables==3.1.0
mock
